+++ tomcat/src/main/java/org/apache/catalina/Manager.java
@@ -1,7 +1,7 @@
 package org.apache.catalina;
 
 import java.io.IOException;
-import org.apache.coyote.session.Session;
+import org.apache.coyote.http.session.Session;
 
 /**
  * A <b>Manager</b> manages the pool of Sessions that are associated with a
