+++ tomcat/src/main/java/org/apache/coyote/controller/Handler.java
@@ -1,0 +1,8 @@
+package org.apache.coyote.controller;
+
+import org.apache.coyote.http.request.Request;
+import org.apache.coyote.http.response.Response;
+
+public interface Handler {
+    Response handle(Request request);
+}
