+++ tomcat/src/main/java/org/apache/coyote/controller/LoginHandler.java
@@ -1,13 +1,13 @@
-package org.apache.coyote.handler;
+package org.apache.coyote.controller;
 
 import com.techcourse.db.InMemoryUserRepository;
 import com.techcourse.model.User;
 import java.util.Map;
-import org.apache.coyote.common.Header;
-import org.apache.coyote.common.Request;
-import org.apache.coyote.common.RequestBody;
-import org.apache.coyote.common.Response;
-import org.apache.coyote.common.StatusCode;
+import org.apache.coyote.http.Header;
+import org.apache.coyote.http.StatusCode;
+import org.apache.coyote.http.request.Request;
+import org.apache.coyote.http.request.RequestBody;
+import org.apache.coyote.http.response.Response;
 
 public class LoginHandler implements Handler {
