+++ tomcat/src/main/java/org/apache/coyote/controller/RegisterHandler.java
@@ -1,14 +1,14 @@
-package org.apache.coyote.handler;
+package org.apache.coyote.controller;
 
-import static org.apache.coyote.common.StatusCode.BAD_REQUEST;
+import static org.apache.coyote.http.StatusCode.BAD_REQUEST;
 
 import com.techcourse.db.InMemoryUserRepository;
 import com.techcourse.model.User;
 import java.util.Map;
-import org.apache.coyote.common.Header;
-import org.apache.coyote.common.Request;
-import org.apache.coyote.common.Response;
-import org.apache.coyote.common.StatusCode;
+import org.apache.coyote.http.Header;
+import org.apache.coyote.http.StatusCode;
+import org.apache.coyote.http.request.Request;
+import org.apache.coyote.http.response.Response;
 
 public class RegisterHandler implements Handler {
