--- tomcat/src/main/java/org/apache/coyote/handler/Handler.java
@@ -1,8 +1,0 @@
-package org.apache.coyote.handler;
-
-import org.apache.coyote.common.Request;
-import org.apache.coyote.common.Response;
-
-public interface Handler {
-    Response handle(Request request);
-}
