+++ tomcat/src/main/java/org/apache/coyote/http/Cookie.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.common;
+package org.apache.coyote.http;
 
 import java.util.Arrays;
 import java.util.HashMap;
