+++ tomcat/src/main/java/org/apache/coyote/http/StatusCode.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.common;
+package org.apache.coyote.http;
 
 public enum StatusCode {
     OK(200, "OK"),
@@ -7,7 +7,8 @@ public enum StatusCode {
     UNAUTHORIZED(401, "Unauthorized"),
     NOT_FOUND(404, "Not Found"),
     METHOD_NOT_ALLOWED(405, "Method Not Allowed"),
-    INTERNAL_SERVER_ERROR(500, "Internal Server Error");
+    INTERNAL_SERVER_ERROR(500, "Internal Server Error"),
+    ;
 
     private final int code;
     private final String message;
