+++ tomcat/src/main/java/org/apache/coyote/http/request/RequestBody.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.common;
+package org.apache.coyote.http.request;
 
 import java.util.Arrays;
 import java.util.HashMap;
